backend/requirements.txt
# Core Framework
fastapi>=0.110.0,<0.120.0
uvicorn[standard]>=0.25.0,<0.35.0
starlette>=0.37.0,<0.46.0
pydantic>=2.0.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
sentry-sdk[fastapi]>=2.0.0,<3.0.0

# Database
sqlalchemy>=2.0.0,<3.0.0
asyncpg>=0.29.0,<0.31.0
psycopg2-binary>=2.9.0,<3.0.0
alembic>=1.13.0,<2.0.0

# Redis
redis[hiredis]>=5.0.0,<6.0.0

# HTTP Clients
httpx>=0.27.0,<0.29.0
aiohttp>=3.9.0,<4.0.0
requests>=2.31.0,<3.0.0

# LLM Providers
openai>=1.0.0,<2.0.0
google-generativeai>=0.8.0,<1.0.0
mistralai>=1.0.0,<2.0.0
groq>=0.10.0,<1.0.0

# PDF Processing
PyMuPDF>=1.23.0,<2.0.0
pdfplumber>=0.10.0,<0.12.0
pdfminer.six>=20231228,<20241001
pypandoc>=1.12.0,<2.0.0
python-docx==1.1.0

# Utilities
python-multipart>=0.0.6,<0.1.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
pyyaml>=6.0.0,<7.0.0
python-magic>=0.4.27,<0.5.0
openpyxl>=3.1.0,<4.0.0

# Testing
pytest>=7.4.0,<9.0.0
pytest-asyncio>=0.21.0,<1.0.0

# Code Quality
black>=23.0.0,<25.0.0
flake8>=6.0.0,<9.0.0
mypy>=1.5.0,<2.0.0
isort>=5.12.0,<6.0.0

# Data Processing
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<3.0.0

# Extras
rich>=13.0.0,<15.0.0
tenacity>=8.2.0,<10.0.0
websockets>=12.0.0,<16.0.0
aiofiles>=23.0.0,<25.0.0
